requirements.txt
fastapi
uvicorn
groq
python-dotenv
pymongo
schedule
pyttsx3